requirements.txt
teachable-machine
opencv-python
easyocr
pyserial
